Change mechanics of how the input word is chosen
Some terminals, such as Fish, insert text at the end of the line the
user is currently typing on. Because of this, the previous method of
finding the last word input by the user, which assumed it was the last
word of the buffer, was yielding inconsistent results (e.g. completing
the time at the end of the line).

The current approach takes advantage of the fact the cursor stays in its
current position when the tmux enters copy-mode.

Therefore, this approach does the following:

1. Enters copy mode
2. Move the cursor to the left
3. Starts selection
4. Moves to the previous word
5. Copies the selection
6. Cancel copy mode

With this, I was able to reliably complete words, even in Fish, which
has the time at the end of the line.